requirements.txt

gunicorn
flask
python-dotenv
requests
waitress
